Fix T70725: Frame Symmetry Stroke
Frame Selected centers around the last valid stroke.
When Symmetry is enabled the last mirrored `location` was added to the
`average_stroke_accum` in stead of the original stroke location.

This patch will add the `true_location` to the `average_stroke_accum`.
This contains the original stroke location.

Issue happened in Vertex and Weight paint.

Reviewed By:
	Pablo Dobarro

Differential Revision: https://developer.blender.org/D6161